+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add `NoSolutionFound` exception
 - Add `open` keyword argument to `start_meshcat_visualizer`
+- Introduce *constraints* to enforce tasks as QP equality constraints
 - examples: Detail that quadprog works best on Stretch examples
 - examples: Installation instruction for loop-rate-limiters
 - examples: Switch to DAQP solvers in examples where it works well