+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Task: `JointVelocityTask` for tracking a reference joint-velocity trajectory
+
 ## [3.3.0] - 2025-05-15
 
 ### Added