+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Task: `JointVelocityTask` for tracking a reference joint-velocity trajectory
 
+### Removed
+
+- Remove unused attribute of `SelfCollisionBarrier`
+
 ## [3.3.0] - 2025-05-15
 
 ### Added