+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.4.0] - 2025-07-29
+
 ### Added
 
 - Task: `JointVelocityTask` for tracking a reference joint-velocity trajectory
@@ -397,7 +399,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Python package infrastructure
 
-[unreleased]: https://github.com/qpsolvers/qpsolvers/compare/v3.3.0...HEAD
+[unreleased]: https://github.com/qpsolvers/qpsolvers/compare/v3.4.0...HEAD
+[3.4.0]: https://github.com/qpsolvers/qpsolvers/compare/v3.3.0...v3.4.0
 [3.3.0]: https://github.com/qpsolvers/qpsolvers/compare/v3.2.0...v3.3.0
 [3.2.0]: https://github.com/qpsolvers/qpsolvers/compare/v3.1.0...v3.2.0
 [3.1.0]: https://github.com/qpsolvers/qpsolvers/compare/v3.0.0...v3.1.0